# Add DecodeDraw to allow custom image types