Closes #2