@bbrks 2020-07-29
Hey, thanks for the contribution!

The detailed writeup in the issue makes sense to me, as well as the cursory glance over the PR.
I'll have a play around with this a little bit in my day tomorrow and merge if all looks good 👍