@bbrks 2020-08-05
Sorry, this is taking a little longer than expected. I ran into an issue when testing this where the new interface-based `image.Set` is causing a huge increase in allocations for the default Decode method (for NRGBA images).

I can work around it by doing a type check to switch to the faster paths for known image types, but want to be sure I've not screwed anything up in the process of switching between alpha and non-alpha premultiplied colors when doing so.

```golang
16:50 $ benchcmp bench-master.txt bench-new.txt
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                          old ns/op     new ns/op     delta
BenchmarkDecode/LFE.@D9F01_2%L%MIVD*9Goe-;WB-8     820101        815481        -0.56%
BenchmarkDecode/LNAdApj[00aymkj[TKay9}ay-Sj[-8     750371        761578        +1.49%
BenchmarkDecode/LNMF%n00%#MwS|WCWEM{R*bbWBbH-8     816556        823240        +0.82%
BenchmarkDecode/KJG8_@Dgx]_4V?xuyE%NRj-8           687708        699058        +1.65%

benchmark                                          old allocs     new allocs     delta
BenchmarkDecode/LFE.@D9F01_2%L%MIVD*9Goe-;WB-8     3              1027           +34133.33%
BenchmarkDecode/LNAdApj[00aymkj[TKay9}ay-Sj[-8     3              1027           +34133.33%
BenchmarkDecode/LNMF%n00%#MwS|WCWEM{R*bbWBbH-8     3              1027           +34133.33%
BenchmarkDecode/KJG8_@Dgx]_4V?xuyE%NRj-8           3              1027           +34133.33%

benchmark                                          old bytes     new bytes     delta
BenchmarkDecode/LFE.@D9F01_2%L%MIVD*9Goe-;WB-8     4448          8544          +92.09%
BenchmarkDecode/LNAdApj[00aymkj[TKay9}ay-Sj[-8     4448          8544          +92.09%
BenchmarkDecode/LNMF%n00%#MwS|WCWEM{R*bbWBbH-8     4448          8544          +92.09%
BenchmarkDecode/KJG8_@Dgx]_4V?xuyE%NRj-8           4384          8480          +93.43%
```