@diamondburned 2020-08-05
That's an interesting observation. I'm guessing it has to do with the type assertion [here](https://golang.org/src/image/color/color.go#L193), which possibly creates a copy of `color.NRGBA`.

I think a fast path approach is reasonable. Something like this, I assume:

```go
setter, isNRGBA := dst.(*image.NRGBA)

for .. {
    for .. {
        if isNRGBA {
            setter.SetNRGBA(x, y, color.NRGBA{})
        } else {
            dst.Set(x, y, color.NRGBA{})
        }
    }
}
```