@bbrks 2020-08-05
I think this is going to end up being just as alloc heavy when using an RGBA image instead of an NRGBA image. I've played around a little and got a fast path low-alloc approach for both cases. Only downside that I can see is the type switch is inside the loop, but the benchmark doesn't show any impact.

What do you think?

```golang
type drawImageNRGBA interface {
	SetNRGBA(x, y int, c color.NRGBA)
}

type drawImageRGBA interface {
	SetRGBA(x, y int, c color.RGBA)
}
```

```golang
for .. {
    for .. {

			switch d := dst.(type) {
			case *image.NRGBA:
				d.SetNRGBA(x, y, color.NRGBA{
					R: uint8(linearTosRGB(r)),
					G: uint8(linearTosRGB(g)),
					B: uint8(linearTosRGB(b)),
					A: 255,
				})
			case *image.RGBA:
				d.SetRGBA(x, y, color.RGBA{
					R: uint8(linearTosRGB(r)),
					G: uint8(linearTosRGB(g)),
					B: uint8(linearTosRGB(b)),
					A: 255,
				})
			default:
				d.Set(x, y, color.NRGBA{
					R: uint8(linearTosRGB(r)),
					G: uint8(linearTosRGB(g)),
					B: uint8(linearTosRGB(b)),
					A: 255,
				})
			}

```

```
22:01 $ benchstat orig.txt pr.txt pr-switch2.txt
name \ time/op                             orig.txt     pr.txt        pr-switch2.txt
Decode/LFE.@D9F01_2%L%MIVD*9Goe-;WB-8       874µs ± 7%    825µs ± 1%      795µs ± 1%
Decode/LNAdApj[00aymkj[TKay9}ay-Sj[-8       801µs ± 3%    765µs ± 1%      733µs ± 1%
Decode/LNMF%n00%#MwS|WCWEM{R*bbWBbH-8       881µs ±10%    831µs ± 0%      798µs ± 1%
Decode/KJG8_@Dgx]_4V?xuyE%NRj-8             756µs ±11%    702µs ± 1%      669µs ± 1%
DecodeDraw/LFE.@D9F01_2%L%MIVD*9Goe-;WB-8                 851µs ± 1%      813µs ± 3%
DecodeDraw/LNAdApj[00aymkj[TKay9}ay-Sj[-8                 787µs ± 0%      742µs ± 3%
DecodeDraw/LNMF%n00%#MwS|WCWEM{R*bbWBbH-8                 852µs ± 0%      797µs ± 1%
DecodeDraw/KJG8_@Dgx]_4V?xuyE%NRj-8                       726µs ± 1%      668µs ± 0%

name \ alloc/op                            orig.txt     pr.txt        pr-switch2.txt
Decode/LFE.@D9F01_2%L%MIVD*9Goe-;WB-8      4.45kB ± 0%   8.54kB ± 0%     4.45kB ± 0%
Decode/LNAdApj[00aymkj[TKay9}ay-Sj[-8      4.45kB ± 0%   8.54kB ± 0%     4.45kB ± 0%
Decode/LNMF%n00%#MwS|WCWEM{R*bbWBbH-8      4.45kB ± 0%   8.54kB ± 0%     4.45kB ± 0%
Decode/KJG8_@Dgx]_4V?xuyE%NRj-8            4.38kB ± 0%   8.48kB ± 0%     4.38kB ± 0%
DecodeDraw/LFE.@D9F01_2%L%MIVD*9Goe-;WB-8                8.48kB ± 0%     0.29kB ± 0%
DecodeDraw/LNAdApj[00aymkj[TKay9}ay-Sj[-8                8.48kB ± 0%     0.29kB ± 0%
DecodeDraw/LNMF%n00%#MwS|WCWEM{R*bbWBbH-8                8.48kB ± 0%     0.29kB ± 0%
DecodeDraw/KJG8_@Dgx]_4V?xuyE%NRj-8                      8.42kB ± 0%     0.23kB ± 0%

name \ allocs/op                           orig.txt     pr.txt        pr-switch2.txt
Decode/LFE.@D9F01_2%L%MIVD*9Goe-;WB-8        3.00 ± 0%  1027.00 ± 0%       3.00 ± 0%
Decode/LNAdApj[00aymkj[TKay9}ay-Sj[-8        3.00 ± 0%  1027.00 ± 0%       3.00 ± 0%
Decode/LNMF%n00%#MwS|WCWEM{R*bbWBbH-8        3.00 ± 0%  1027.00 ± 0%       3.00 ± 0%
Decode/KJG8_@Dgx]_4V?xuyE%NRj-8              3.00 ± 0%  1027.00 ± 0%       3.00 ± 0%
DecodeDraw/LFE.@D9F01_2%L%MIVD*9Goe-;WB-8                 2.05k ± 0%      0.00k ± 0%
DecodeDraw/LNAdApj[00aymkj[TKay9}ay-Sj[-8                 2.05k ± 0%      0.00k ± 0%
DecodeDraw/LNMF%n00%#MwS|WCWEM{R*bbWBbH-8                 2.05k ± 0%      0.00k ± 0%
DecodeDraw/KJG8_@Dgx]_4V?xuyE%NRj-8                       2.05k ± 0%      0.00k ± 0%
```