@diamondburned 2020-08-05
I think the color should be created outside the type switch to make it shorter. Other than that, I think it's fine.